requirements.txt
#djangorestframework==2.3.13    # It's not necessary since Taiga 1.7
CairoSVG==2.0.3
Django==1.11.2
Markdown==2.6.8
Pillow==4.1.1
PyJWT==1.5.0
Unidecode==0.4.20
amqp==2.1.4
asana==0.6.5
bleach==2.0.0
celery==4.0.2
cryptography==1.9
cssutils==1.0.2
diff-match-patch==20121119
django-ipware==1.1.6
django-jinja==2.3.1
django-picklefield==0.3.2
django-sampledatahelper==0.4.1
django-sites==0.9
django-sr==0.0.4
djmail==1.0.1
easy-thumbnails==2.4.1
fn==0.4.3
git+https://github.com/Xof/django-pglocks.git
gunicorn==19.7.1
jinja2==2.9.6
lxml==3.8.0
netaddr==0.7.19
premailer==3.0.1
psd-tools==1.4
psycopg2==2.7.1
pygments==2.2.0
pyjwkest==1.3.2
python-dateutil==2.6.0
python-magic==0.4.13
pytz==2017.2
raven==6.1.0
redis==2.10.5
requests-oauthlib==0.8.0
requests==2.17.3
serpy==0.1.1
six==1.10.0
webcolors==1.7
django-push-notifications==1.6.0
